Renaming test libs to extension
